+++ mujoco_ros/include/mujoco_ros/mujoco_env.h
@@ -46,32 +46,37 @@ namespace MujocoSim {
 
 namespace environments {
 
-static std::map<mjData *, MujocoEnvPtr> env_map_;
-
 /**
- * @brief Adds a new entry to the env_map_, assigning a MujocoEnvPtr to a unique mjData pointer.
+ * @brief Assigns an mjData pointer to a MujocoEnv.
  *
  * @param[in] data mjData pointer used as key element in the map.
  * @param[in] env MujocoEnvPtr assigned to the mjData pointer.
  */
 void assignData(mjData *data, MujocoEnvPtr env);
 
+[[deprecated("Multienv is no longer supported. Use MujocoSim::detail::main_env_ to get the only available env "
+             "instead.")]]
 /**
  * @brief Retrieve the MujocoEnvPtr assigned to the given mjData pointer from the env_map_.
  *
  * @param[in] data mjData pointer used as key.
  * @return MujocoEnvPtr if `data` is in the map, otherwise nullptr is returned.
  */
-MujocoEnvPtr getEnv(mjData *data);
+MujocoEnvPtr
+getEnv(mjData *data);
 
+[[deprecated("Multienv is no longer supported. Use MujocoSim::detail::main_env_ to get the only available env "
+             "instead.")]]
 /**
  * @brief Retreive a MujocoEnvPtr by id.
  *
  * @param[in] id id to search for.
  * @return MujocoEnvPtr if an env with that id exists, otherwise nullptr is returned.
  */
-MujocoEnvPtr getEnvById(uint id);
+MujocoEnvPtr
+getEnvById(uint id);
 
+[[deprecated("Multienv is no longer supported. Unregistering is no longer necessary.")]]
 /**
  * @brief Unregisters an mjData/MujocoEnvPtr entry in the static map resolved by mjData pointer.
  * This is necessary to deallocate the environment and its plugins, once it is not needed anymore, to greacefully shut
@@ -81,6 +86,7 @@ MujocoEnvPtr getEnvById(uint id);
  */
 void unregisterEnv(mjData *data);
 
+[[deprecated("Multienv is no longer supported. Unregistering is no longer necessary.")]]
 /**
  * @brief Unregisters an mjData/MujocoEnvPtr entry in the static map resolved by environment id.
  * This is necessary to deallocate the environment and its plugins, once it is not needed anymore, to greacefully shut